// src/types/routeParams.ts
export enum type {
  single = "single",
  joint = "joint",
}
export enum period {
  current = "current",
  previous = "previous",
}
export enum person {
  first = "1",
  second = "0",
}

export enum idNumber {
  "one" = "1",
  "two" = "2",
}
export enum section {
  employmentDetails = "employmentDetails",
  personalDetails = "personalDetails",
  incomeExpense = "incomeExpense",
}